# openapi.yaml
openapi: 3.1.0
info:
  title: Mahshar V2 Agent API
  version: 2.1.0
  description: |
    Public machine interface for discovering and purchasing proxied APIs with
    Circle Gateway x402 v2 on Arc Mainnet (chain 5042, CAIP-2 eip155:5042).
    Seller upstream URLs and credentials are never part of this contract.
servers:
  - url: https://mahshar.xyz
...
paths:
...
  /api/agent/discover:
    get:
      tags: [Discovery]
      operationId: discoverApis
      summary: Discover active listings and the public payment contract
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, maximum: 10000, default: 0 } }
      responses:
        '200':
          description: Agent catalog
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Discovery' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /api/calls/last-response:
    get:
      tags: [Purchase access]
      operationId: retrievePurchasedResponse
      summary: Retrieve the private response for the exact authorized purchase
      description: |
        Send `x-mahshar-purchase-access` with the purchase-scoped capability.
        A legacy purchase may exchange one one-use `x-mahshar-authorization`
        wallet proof for a capability. `buyer_wallet` alone is never ownership.
      parameters:
        - { name: api_id, in: query, required: true, schema: { type: string } }
        - { name: buyer_wallet, in: query, required: true, schema: { type: string } }
        - { name: x-mahshar-purchase-access, in: header, required: false, schema: { type: string } }
        - { name: x-mahshar-authorization, in: header, required: false, schema: { type: string } }
      responses:
        '200':
          description: Stored response and optionally a newly issued capability
          content:
            application/json:
              schema:
                type: object
                properties:
                  response_body: {}
                  purchase_access_token: { type: string }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /api/payments/reconcile:
    post:
      tags: [Recovery]
      operationId: reconcilePaymentAccounting
      summary: Reconcile accounting; never settles or replays upstream delivery
      parameters:
        - { name: x-mahshar-authorization, in: header, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [attemptId]
              properties:
                attemptId: { type: string, format: uuid }
      responses:
        '200': { description: Accounting complete }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/DeliveryConflict' }
        '503': { $ref: '#/components/responses/Error' }
components:
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        message: { type: string }
        attemptId: { type: string }
        delivery_state: { type: string }
        retryable: { type: boolean }
    AgentListing:
      type: object
      required: [id, name, method, proxy_url, proxy_style, request, response]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        category: { type: string }
        price_per_call_usdc: { type: number }
        payment_model: { const: x402-pay-per-call }
        method: { type: string, enum: [GET, POST, PUT, DELETE] }
        proxy_style: { type: string, enum: [path, envelope] }
        proxy_url: { type: string, format: uri }
        auth:
          type: object
          properties:
            type: { enum: [public, apikey, bearer, queryparam] }
            injected_by: { const: mahshar }
            seller_credentials_exposed: { const: false }
        request:
          type: object
          properties:
            outer_method: { type: string }
            content_type: { type: [string, 'null'] }
            body: { type: object }
            dynamic_path: { type: object }
            path_parameters: { type: array }
            query_parameters: { type: array }
            incoming_headers: { type: object }
        response: { type: object }
    Discovery:
      type: object
      required: [marketplace, contract_version, openapi_url, network, payment_protocol, payment_recipient, payment_domain, apis]
      properties:
        marketplace: { const: Mahshar }
        contract_version: { type: string }
        openapi_url: { type: string, format: uri }
        network: { const: 'eip155:5042' }
        payment_protocol: { const: x402 }
        payment_recipient:
          type: string
          description: Expected x402 payTo address. Clients must compare it with the challenge before signing.
        payment_domain:
          type: object
          description: Expected EIP-712 domain. Clients must compare it with the challenge before signing.
        networks: { type: array }
        payment_signature_schema: { type: object }
        recovery: { type: object }
        rate_limits: { type: object }
        pagination: { type: object }
        apis:
          type: array
          items: { $ref: '#/components/schemas/AgentListing' }
  responses:
...
    Error:
      description: Machine-readable error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    DeliveryConflict:
      description: Payment accounted; delivery pending, final, or unknown. Do not authorize a second payment.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/Error' }
              - type: object
                properties:
                  error: { const: rate_limited }
                  retry_after_seconds: { type: integer }